Refactor search functions to remove unused error checks and improve pattern length validation

+++ crates/fuse-rs/src/search/bitmap/create_pattern_alphabet.rs
@@ -11,7 +11,7 @@
 /// # Returns
 ///
 /// A HashMap where keys are characters and values are bitmasks
-pub fn create_pattern_alphabet(pattern: &str) -> std::collections::HashMap<char, u64> {
+pub fn create_pattern_alphabet(pattern: &str) -> std::collections::HashMap<char, u32> {
     let mut mask = std::collections::HashMap::new();
     let len = pattern.len();
